aruco-tracker-client/pyproject.toml
[tool.poetry]
name = "aruco-tracker-client"
version = "0.1.0"
description = ""
authors = ["Heng <31088159+shiukaheng@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.20.1"
opencv-contrib-python = "^4.5.1"
Pillow = "^8.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
